code_agent/requirements.txt
# Core dependencies for the Python agent
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
